Assertion and self-testing support for TF
This adds support for assertions to the
Test Framework so that a single test
can fail on an any number of conditions
whilst still only being recorded as a
single failure.

This is added in the form of test_assert
and changes to test_start and test_finish.
Note that backwards compatibility for simple
tests passing 'expected' and 'actual' values
to test_finish has been retained.

In addition, this patch also adds
'test_test_framework' as a module.
This runs a bunch of tests  with and
against its own functions to try to detect
issues with the framework itself.