Refactor cslib_menu
This should be a largely non-functional commit,
albeit a large one. The intent is to make the
module's logic a bit easier to understand by
splittings a lot of the functionality
out into internal named routines. I've also
renamed some variables and routines, and
improved the common elements.

Choice building was also restructured so that
array and simple menus can share more page logic.