Return 'filter' support to cslib_array via routines
We removed the filter support from cslib_menu to
simplify the module, thinking we could rely on
the 'filter' routine from cslib_array.
Unfortunately that was an oversight. When you use
filter in cslib_array, you're creating a new array,
with new element indexes.

This means that the specific use-case of using cslib_menu's
return to index into another attribute of the array is lost.

E.g. with cslib's test suite:
*gosub_scene cslib_menu build_array_filter "module" ...
${module_log[cslib_ret}

The above is no longer possible with cslib_array filter.

This patch returns that functionality to cslib_menu via
a routine implementation (passing a scene+label), which
has a few advantages:
- It matches up with cslib_array implementations
- It simplifies the code in cslib_menu (over the original)
- It doesn't require us to mandate a particular arr attr format
- The parameter list isn't as indecipherable

On the negatives side, the performance is arguably worse,
but it seems like a reasonable trade-off. You can still use
cslib_array filter as well.